Arreglar esto

+++ app/main/routes.py
@@ -25,18 +25,25 @@ def index():
 def edit_profile():
     form = EditProfileForm(current_user.username)
     if form.validate_on_submit():
-        tmp_dir = os.path.join(upload_temp_dir, request.form.get('avatar'))
-        tmp_file = os.listdir(tmp_dir)[0]
-        shutil.move(os.path.join(tmp_dir, tmp_file), os.path.join(Config.UPLOADED_AVATARS_DEST, tmp_file))
-        print('MOVIENDO: '+ os.path.join(Config.UPLOADED_AVATARS_DEST, tmp_file))
-        os.rmdir(tmp_dir)
-        if current_user.avatar:
-            try:
-                os.remove(os.path.join(Config.UPLOADED_AVATARS_DEST, current_user.avatar))
-            except:
-                print('Error al eliminar el archivo '+ os.path.join(Config.UPLOADED_AVATARS_DEST, current_user.avatar))
-        current_user.avatar = tmp_file
+        if current_user.avatar != request.form.get('avatar'):
+            print('avatar: '+request.form.get('avatar'))
+            print('upload_temp_dir: ' + upload_temp_dir)
+            tmp_dir = os.path.join(upload_temp_dir, request.form.get('avatar'))
+            print('tmp_dir: '+tmp_dir)
+            tmp_file = os.listdir(tmp_dir)[0]
+            print('tmp_file: '+tmp_file)
+            shutil.move(os.path.join(tmp_dir, tmp_file), os.path.join(Config.UPLOADED_AVATARS_DEST, tmp_file))
+            print('MOVIENDO: '+ os.path.join(Config.UPLOADED_AVATARS_DEST, tmp_file))
+            os.rmdir(tmp_dir)
+            if current_user.avatar:
+                try:
+                    os.remove(os.path.join(Config.UPLOADED_AVATARS_DEST, current_user.avatar))
+                except:
+                    print('Error al eliminar el archivo '+ os.path.join(Config.UPLOADED_AVATARS_DEST, current_user.avatar))
+                    current_user.avatar = tmp_file
         current_user.username = form.username.data
+        current_user.first_name = form.first_name.data
+        current_user.last_name = form.last_name.data
         db.session.commit()
         flash('Your changes have been saved.')
         return redirect(url_for('main.edit_profile'))
